LayoutTests/js/promise-isPromise-expected.txt

IDL promises should work with Promise.isPromise

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Promise.isPromise((new Blob).text()) is true
PASS Promise.isPromise((new Blob).arrayBuffer()) is true
PASS Promise.isPromise((new Blob).bytes()) is true
PASS Promise.isPromise(document.createElement("img").decode()) is true
PASS Promise.isPromise((new ReadableStream).cancel()) is true
PASS Promise.isPromise((new ReadableStream).pipeTo(new WritableStream)) is true
PASS Promise.isPromise((new WritableStream).abort()) is true
PASS Promise.isPromise((new WritableStream).close()) is true
PASS Promise.isPromise(event.promise) is true
PASS successfullyParsed is true

TEST COMPLETE
